refactor(search): GlobalSearchModal sous 300 lignes — éclatement complet (H13)
726 -> 171 l. Logique (debounce, searchService, navigation clavier, historique) -> useGlobalSearch.js ; sections -> SearchHeader/SearchResultsList/SearchEmptyState/SearchFooter. API (modelValue, update:modelValue/result-selected) inchangee — DashboardLayout intact. CSS verbatim. Tests parite verts + nouveaux. Dette preexistante conservee (CSS .spinner mort, console.log debug).
